pyproject.toml
[tool.poetry]
name = "biblio"
version = "0.1.0"
description = "Database of books I read."
authors = ["Iñaki Silanes <isilanes@gmail.com>"]
license = "GPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1"
psycopg2-binary = "^2.9.4"
plotly = "^5.10.0"
djangorestframework = "^3.14.0"

[tool.poetry.group.dev.dependencies]
radon = "^5.1.0"
coverage = "^6.5.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
